update the project
